Fill textviews in transfer confirmation , closed #314 , create transfer repository , closed #313 , call start transfer request , closed #308

+++ app/src/main/java/de/netalic/falcon/data/model/Action.java
@@ -14,7 +14,7 @@ public class Action {
     private String mType;
 
     @SerializedName("amount")
-    private double mAmount;
+    private float mAmount;
 
     @SerializedName("currencyCode")
     private String mCurrencyCode;
